Initial take on a Vagrantfile.
Started from what you get with `vagrant init` and tried to customize
some. The yum command doesn't seem to work right as the package is not
installed by default as the machine comes up.

Remember to use vagrant ssh to get into the machine and vagrant halt
to shut it down.